generalizing all to 'entity', added account deletion explanation window and generalizing sort filter
