+++ src/main/java/com/tim23/fishnchill/reservation/dto/CottageOwnerCottageReservationDto.java
@@ -25,7 +25,7 @@ public class CottageOwnerCottageReservationDto {
 
     private LocalDateTime reservationEnd;
 
-    private CottageDto cottage;
+    private CottageDto entity;
 
     private ClientDto client;
