+++ web/pages/reservations/past.jsx
@@ -3,6 +3,7 @@ import { useState, useEffect } from 'react'
 import reservationService from '../../services/reservation'
 import clientService from '../../services/client'
 import ratingService from '../../services/rating'
+import dateUtils from 'utils/dateUtils'
 
 const ReservationHistory = () => {
   const [reservations, setReservations] = useState([])
@@ -10,11 +11,14 @@ const ReservationHistory = () => {
   const [beginingRatings, setBeginingRatings] = useState([])
   const [statusMessage, setStatusMessage] = useState('')
   const [submitStatusMessage, setSubmitStatusMessage] = useState('')
+  const [sortFilterItems, setSortFilterItems] = useState(['Name', 'Duration', 'Price'])
+  const [ratings, setRatings] = useState([])
 
   useEffect(() => {
     clientService
       .getRatings()
       .then((gotRatings) => {
+        setRatings(gotRatings)
         reservationService
           .getAllPastCottageReservationsForClient()
           .then((gotReservations) => {
@@ -22,7 +26,7 @@ const ReservationHistory = () => {
             let ratingss = new Array(gotReservations.length).fill(0)
             gotReservations.forEach((reservation, index) => {
               gotRatings.forEach((rating) => {
-                if (reservation.cottage.id === rating.entity.id) ratingss[index] = rating.rating
+                if (reservation.entity.id === rating.entity.id) ratingss[index] = rating.rating
               })
             })
             setRatingsProp(ratingss)
@@ -64,6 +68,42 @@ const ReservationHistory = () => {
         .catch((err) => console.log(err))
   }
 
+  const handleSortFilterChange = (e) => {
+    let newReservations
+    switch (e.target.value) {
+      case 'name': {
+        newReservations = [...reservations.sort((a, b) => a.entity.name.localeCompare(b.name))]
+        break
+      }
+      case 'duration': {
+        newReservations = [
+          ...reservations.sort(
+            (a, b) =>
+              dateUtils.daysBetween(new Date(b.reservationEnd), new Date(b.reservationStart)) -
+              dateUtils.daysBetween(new Date(a.reservationEnd), new Date(a.reservationStart))
+          ),
+        ]
+        break
+      }
+      case 'price': {
+        newReservations = [...reservations.sort((a, b) => a.price - b.price)]
+        break
+      }
+      default: {
+        break
+      }
+    }
+    setReservations(newReservations)
+    let ratingss = new Array(newReservations.length).fill(0)
+    newReservations.forEach((reservation, index) => {
+      ratings.forEach((rating) => {
+        if (reservation.entity.id === rating.entity.id) ratingss[index] = rating.rating
+      })
+    })
+    setRatingsProp(ratingss)
+    setBeginingRatings(ratingss)
+  }
+
   if (ratingsProp.length !== 0)
     return (
       <PastReservations
@@ -75,6 +115,8 @@ const ReservationHistory = () => {
         statusMessage={statusMessage}
         submitResponse={submitResponse}
         submitStatusMessage={submitStatusMessage}
+        handleSortFilterChange={handleSortFilterChange}
+        sortFilterItems={sortFilterItems}
       />
     )
   else return <></>
